project/requirements.txt
flask==3.1.0
flask-cors==5.0.0
httpx==0.27.0
python-dotenv==1.0.1
datasets==3.5.0
